refactor(getCurrentPeriod): rename method to getCurrentEntity and made it universal
