+++ src/assets/utils/getCurrentEntity.js
@@ -1,0 +1,5 @@
+const getCurrentEntity = (array, id) => {
+  return array.find((entity) => entity.id === id);
+};
+
+export default getCurrentEntity;
