--- src/assets/utils/getCurrentPeriod.js
@@ -1,5 +1,0 @@
-const getCurrentPeriod = (periods, id) => {
-  return periods.find((period) => period.id === id);
-};
-
-export default getCurrentPeriod;
